Add validation for email address when creating new users

+++ echobot.py
@@ -214,28 +214,32 @@ async def toggle_input(update: Update, context: ContextTypes.DEFAULT_TYPE) -> st
 
 
 async def adding_user(update: Update, context: ContextTypes.DEFAULT_TYPE) -> str:
-    try:
-        uuid_ = str(uuid.uuid4())
-        email = context.user_data[ATTRIBUTES][EMAIL]
-        persistent = context.user_data[ATTRIBUTES][PERSISTENT]
+    if not (email:=context.user_data[ATTRIBUTES].get(EMAIL)):
+        await update.callback_query.answer("Email address cannot be empty.")
+        return
 
-        User.create(email=email, persistent=persistent)
-        add_vless_user(client=XRAY_CTL.hs_client, uuid=uuid_, level=0, in_tag="vless", email=email)
+    if User.select().where(User.email==email).exists():
+        await update.callback_query.answer("This email address is already in use by another user.")
+        return
 
-        if persistent:
-            vless_inbound, = [inbound for inbound in xray_config["inbounds"] if inbound["tag"] == "vless"]
-            vless_inbound["settings"]["clients"].append({"id": uuid_, "email": email})
+    persistent = context.user_data[ATTRIBUTES][PERSISTENT]
+    uuid_ = str(uuid.uuid4())
 
-            with open(XRAY_CONFIG_PATH, "w") as handle:
-                json.dump(xray_config, handle, indent=4)
+    User.create(email=email, persistent=persistent)
+    add_vless_user(client=XRAY_CTL.hs_client, uuid=uuid_, level=0, in_tag="vless", email=email)
 
-        uri = (
-            f"vless://{uuid_}@{ADDRESS}:{PORT}?path={urllib.parse.quote_plus(PATH)}"
-            f"&security=tls&encryption=none&type=ws#{REMARKS}-{email}"
-        )
-        text = f"Here is the URI for user \"{email}\":\n\n<code>{html.escape(uri)}</code>"
-    except RpcError as e:
-        text = f"Failed to add VLESS user:\n\n<pre>{html.escape(str(e))}</pre>"
+    if persistent:
+        vless_inbound, = [inbound for inbound in xray_config["inbounds"] if inbound["tag"] == "vless"]
+        vless_inbound["settings"]["clients"].append({"id": uuid_, "email": email})
+
+        with open(XRAY_CONFIG_PATH, "w") as handle:
+            json.dump(xray_config, handle, indent=4)
+
+    uri = (
+        f"vless://{uuid_}@{ADDRESS}:{PORT}?path={urllib.parse.quote_plus(PATH)}"
+        f"&security=tls&encryption=none&type=ws#{REMARKS}-{email}"
+    )
+    text = f"Here is the URI for user \"{email}\":\n\n<code>{html.escape(uri)}</code>"
 
     await update.callback_query.answer()
     await update.callback_query.edit_message_text(text=text, parse_mode=ParseMode.HTML)
